refactor(component): cleanup grouped files, docs, and deleted comments (#2461)
