+++ modules/component/src/core/cd-aware/get-change-detection-handling.ts
@@ -5,8 +5,8 @@ import {
   ɵmarkDirty as markDirty,
 } from '@angular/core';
 
-import { isIvy } from './is-ivy';
-import { hasZone } from './has-zone';
+import { isIvy } from '../utils/is-ivy';
+import { hasZone } from '../utils/has-zone';
 
 export function getChangeDetectionHandler(
   ngZone: NgZone,
@@ -14,19 +14,6 @@ export function getChangeDetectionHandler(
 ): <T>(component?: T) => void {
   if (isIvy()) {
     return hasZone(ngZone) ? markDirty : detectChanges;
-  } else {
-    return hasZone(ngZone)
-      ? cdRef.markForCheck.bind(cdRef)
-      : cdRef.detectChanges.bind(cdRef);
-  }
-}
-
-export function getDetectChanges(
-  ngZone: NgZone,
-  cdRef: ChangeDetectorRef
-): <T>(component?: T) => void {
-  if (isIvy()) {
-    return !hasZone(ngZone) ? detectChanges : markDirty;
   } else {
     return hasZone(ngZone)
       ? cdRef.markForCheck.bind(cdRef)
