+++ modules/component/src/core/cd-aware/index.ts
@@ -1,0 +1,2 @@
+export * from './get-change-detection-handling';
+export * from './cd-aware_creator';
