+++ modules/component/src/core/utils/get-global-this.ts
@@ -1,7 +1,23 @@
-// Returns a reference to global thin
-// - Browser
-// - SSR
-// - Tests
+/**
+ * @description
+ *
+ * This function returns a reference to globalThis in the following environments:
+ * - Browser
+ * - SSR (Server Side Rendering)
+ * - Tests
+ *
+ * @returns {any} - the reference to globalThis in the current environment.
+ *
+ * @usageNotes
+ *
+ * The function can be just imported and used everywhere.
+ *
+ * ```ts
+ * import { getGlobalThis } from `utils/get-global-this`;
+ *
+ * console.log(getGlobalThis());
+ * ```
+ */
 export function getGlobalThis(): any {
   return ((globalThis as any) || (self as any) || (window as any)) as any;
 }
