+++ modules/component/src/core/utils/index.ts
@@ -1,4 +1,3 @@
-export { getChangeDetectionHandler } from './get-change-detection-handling';
 export { getGlobalThis } from './get-global-this';
 export { isIvy } from './is-ivy';
 export { hasZone } from './has-zone';
